+++ R/char_replace.R
@@ -13,7 +13,8 @@
 #' @param rm_dash logical, removes all dash-characters within the string. 
 #' @param to_underscore logical, replaces all whitespace and dash-characters
 #' with underscores.
-#' 
+#' @param trim logical, if whitespaces are trimmed from the start and end.
+#' Is ignored with \code{rm_space == TRUE}.
 #' 
 #' @return a character vector with all non-standard characters replaced by their
 #' standard counterparts
@@ -109,7 +110,8 @@ char_replace <- function(x,
         "ÿ" = "y", 
         "þ" = "th", 
         "ž" = "z", 
-        "Ž" = "Z"))
+        "Ž" = "Z",
+        " " = " "))
   
   if (trim == TRUE) {
     input_processed <- input_processed %>%