+++ R/object_size_in_env.R
@@ -11,6 +11,8 @@
 #' @return Returns a data.table with the size of objects in the given
 #' environment
 #' 
+#' @export
+#' 
 #' @examples
 #' DT <- object_size_in_env(env = environment())
 #' 
@@ -18,6 +20,10 @@ object_size_in_env <- function(env = .GlobalEnv, unit = "Mb") {
   
   # get objects
   tmp_ls <- ls(envir = env)
+  
+  if (length(tmp_ls) == 0 ) {
+    return("no objects in this environment.")
+  }
   
   # create data.table with names and size
   size_table <- data.table::data.table(