+++ tests/testthat/test-statusbar.R
@@ -10,7 +10,14 @@ test_that("error message", {
   
 })
 
+
 test_that("example output is right", {
   expect_output(statusbar(run = 2, max.run = 20, percent.max = 10L),
                 "\\n \\[=         \\]   10\\.00% - 2")
+  
+  expect_output(statusbar(run = "b", max.run = letters[1:3], percent.max = 10L),
+                "\\n \\[=======   \\]   66\\.67% - b")
+  
+  expect_output(statusbar(run = "b", max.run = letters[2], percent.max = 10L),
+                "\\n \\[==========\\]  100\\.00% - b")
 })