+++ tests/testthat/test-tests.R
@@ -1,4 +1,4 @@
-context("check existence of test files")
+context("check tests and scipts")
 
 test_that("if for each function there is an testfile namend 'test-xxx'", {
   # for debugging since .. does not work here
@@ -40,4 +40,45 @@ test_that("if for each function there is an testfile namend 'test-xxx'", {
   
   expect_equal(sum(indx), 0)
   
+})
+
+test_that("exports of funcitons", {
+  # for debugging since .. does not work here
+  # function_files <- list.files("R/", full.names = TRUE)
+  
+  # get functions and tests
+  function_files <- list.files("../../R/", full.names = TRUE)
+  # read files
+  
+  package_files <- c("tests.R", "helfRlein.R")
+  exceptions <- c()
+  if (length(intersect(exceptions, basename(function_files))) != 0) {
+    warning("there are exceptions for files that are not exported: ", 
+            paste0(intersect(exceptions, basename(function_files)),
+                   collapse = ", "))
+  }
+  function_files <- paste0(dirname(function_files), "/",
+                           setdiff(basename(function_files),
+                            c(exceptions, package_files)))
+  
+  
+  files_list <- lapply(function_files, readLines)
+  
+  
+  
+  # check if all fucntions are exported
+  if (length(exceptions) != 0) {
+    warning(paste0("there are still ", length(exceptions),
+                   " export exceptions: ",
+                   paste0(exceptions, collapse = ", ")))
+  }
+  
+  indx <- !grepl(pattern = "#' @export", x = files_list)
+  if (any(indx)) {
+    warning(paste0("there are still ", sum(indx),
+                   " exports missing: ",
+                   paste0(function_files[indx], collapse = ", ")))
+  }
+  
+  expect_equal(sum(indx), 0)
 })